Generative_AI/Langchain/requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
beautifulsoup4
langchain-openai
chromadb
